Complete functionality of reader, with functional changes
Now, on the main loop, the reader behaves as does the writer by reading
in values from the FSR analog input, and by recording a working number
of pressure changes. The reader now differs from the writer in that,
rather than making a post, it fetches the most recent step value from
the server, and then actuates the actuation pin if there are more
shoe steps than walker steps.

Functions have been modularized to support actuating both the designated
actuation pin (for the vibrators and ledStrip), as well as a new
LED to indicate when the shoe is dealing with network overhead and not
listening for sensor values.

Note that if there are no entries within the last 10 min (default,
but changeable in the ?lt part of the GET request), behavior value
fetching is undefined.